requirements.txt
aiomysql
python-dotenv
pyautogui
PySide6
requests
obs-websocket-py